Removed FioMcaLoader plugin. The new ASCII1dProfileLoader plugin allows to load fio files.
